% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_outliers.R
\name{impute_outliers}
\alias{impute_outliers}
\title{Imputes outliers}
\usage{
impute_outliers(x, method = "whisker")
}
\arguments{
\item{x}{A numeric vector}

\item{method}{Method to impute the outliers. Methods that can be used are 'whisker','mean' and 'median'.}
}
\description{
Imputes outliers
}
\examples{
impute_outliers(mpg$hwy,method = 'whisker')
impute_outliers(mpg$hwy,method = 'mean')
}
